refactor: separa os eventos da janela principal
